feat: 对接支付乐自动化充值系统，修复加分漏洞并归档变更记录

+++ backend/schemas/user.py
@@ -42,7 +42,7 @@ class PasswordChange(BaseModel):
     new_password: str
 
 class RechargeApply(BaseModel):
-    money_amount: int
+    money_amount: float
     screenshot_url: Optional[str] = None
 
 class RechargeAudit(BaseModel):
